backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-filter
Pillow
django-cors-headers
python-dotenv
gunicorn
google-auth
google-auth-oauthlib
google-api-python-client
qrcode